Commonise maxwell3d interconnect common code for use by other engines
The compute engine will require most of this for basic functionality.
